Removed global file lock. Adjusted rw-lock. Still passing all regression tests and file extension tests.
The global file lock have been removed and now file system is soley relying on
per-block based read-write lock.

The logic behind the rw-lock was change to address pending writer going back to
sleep after woken up by the last reader. In the previous implementation, a request
from a writer went to pending list if there were any pending readers. However, it
was possible that there could be pending readers when the last reader wake up the
writer. Thus, in that case, writer went back to sleep instead of taking her turns.
Therefore, the new logic was added to properly handle synchronization and fairness
of the rw-lock.