Changed readdir syscall to not assume first two entries are "." and "..". Close directory on exit or upon close(). Now passing all tests.
The previous implementation of readdir syscall made an assumption that the first two entries
must be "." and "..", and skip to the third entry on the first readdir.
However, the use of readdir in a tar file was an issue since the first two entries of a tar
file are not "." and "..". As a result, the contents of the tar file were displayed with two entries
less than they actually were. Therefore, dir_readdir() was modified to skip a entry only if
the name of the entry is either "." or "..".

Additional attempt to avoid memory leakage for directory was made. Now, each process
close current working directory upon exit. In addition, close() syscall now close
open directory from the file descriptor table.

It was observed that all P4 tests (based on P2) are now passing.