Added cache read ahead feature. Added new data structure array_queue. Bug fixed on rw-lock.
The read ahead feature was added to the project. The current approach is to spawn
new kernel thread during file system initialization which would be dedicated for
read ahead and nothing else. Running processes would than enqueue the next sector
to be accessed to the queue, where read-ahead thread would dequeue it and prefetch
the sector to the cache while the requested process is still reading the previous
sector.

The new data structure array queue was implemented to better synchronize the global
read queue. The implementation is based on circular array with consumer/producer
monitor. We provision that this can be reuse for pipe implementation in open ended
project.

Lastly, subtle bug in rw-lock, particularly in write_lock_try_acquire(), was fixed.
The issue was that when write_lock_try_acquire() fail (reutnr false), it did not
release the monitor lock. This was not apparent until now since wirte_lock_try_acqire()
mostly, if not all, succeeded by selecting LRU block in cache.