Bug fixed on write-behind. Added synchronization on open inodes and directories.
The previous write-back implementation contained subtle bug.
The problem happen during cache_flush(), where

    if (write_lock_try_acquire (...) && block->dirty)

it was possible that write_lock_try_acquire() succeed, but block was not dirty.
When that happen, the iteration continued without releasing the write-lock.
Therefore, the order was flipped to always check whether block is dirty first.
In addition, current implementation added checks to make sure the block has
not been evicted while acquiring the lock.

The global inodes_lock was added to protect open inodes list.
The dir_lock was added to each in-memory inode to ensures that names in a
directory are unique.

Lastly, the read syscall was adjusted to reject read on directories.