Flipped cache flush order. Updated several comments.
It was observed that flushing from LRU (back) to MRU (front) is more stable than the opposite
during the periodic write-back (write-behind).
It was possible that block that got flushed gets accessed right after being flushed,
moving to the front of the buffer cache. This effectivly make flush process to
start over from the front of the cache.
With flush order flipped, this is not an issue anymore.