Added open files list. File now keep track of reference count and remain in memory until it reaches 0.
The open files list keep track of files that provide layer of indirection between
file descriptor and inode. This will be use for inheriting parent's file descriptors
to its child.

Also, a open file now keep tracks of reference count and remain in memory until it reaches 0.
Therefore, the close() no longer guaranteed to close the file. This will be useful for dup2()
where it can be implemented by simply increment the reference count.