Hard coding of fd 0 and fd 1 have been removed. Instead, stdin and stdout are now file. Removed file_set_directory().
The code base was tested with examples/dup-stdin.c and examples/dup-stdout.c
Note that all regression tests pass execpt for close-stdout. However, I believe that the
close-stdout is irrelevent as requirement for closing STDOUT is different from P2.
That is, STDOUT should be closeable to be used in pipe.