Added exec2() syscall (Pintos version of execv). Modified cat to read from STDIN when given "-" as a argument.
Support for exec2() syscall, Pintos version of execv, was added to code base.
exec2 replaces the current process image with a new process image.
Note that exec() still work as before, which is ultimately forkexec.

The baisc tests for exec2 (and combination with fork and dup) can be found at
examples/fork-exec.c and example/fork-dup-exec.c. For example, use

    pintos --filesys-size=2 -p ../../examples/fork-dup-exec -a fork-dup-exec -p ../../examples/cat -a cat -- -f -q run 'fork-dup-exec'

to run the test. All regression tests still pass.

Also, examples/cat was modified to read from STDIN when given "-" as a argument.